Remove BotGreet -> move to nsp_wrapper
